requirements.txt
Django==3.2.5
psycopg2==2.9.7
djangorestframework==3.12.4
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.1